@@ -1,39 +1,31 @@
 {
-  "name": "converter-miguelo-dev",
-  "version": "0.1.0",
+  "name": "nuxt-app",
   "private": true,
+  "type": "module",
   "scripts": {
-    "dev": "next dev",
-    "build": "next build",
-    "start": "next start",
-    "lint": "eslint . --ext .ts"
+    "build": "nuxt build",
+    "dev": "nuxt dev",
+    "start": "node .output/server/index.mjs",
+    "generate": "nuxt generate",
+    "preview": "nuxt preview",
+    "postinstall": "nuxt prepare"
   },
   "dependencies": {
-    "file-saver": "^2.0.5",
-    "formik": "^2.4.1",
-    "next": "13.4.4",
-    "react": "18.2.0",
-    "react-dom": "18.2.0",
-    "sharp": "^0.32.1"
+    "@nuxt/ui": "2.20.0",
+    "@nuxtjs/tailwindcss": "^6.12.2",
+    "joi": "^17.13.3",
+    "nuxt": "^3.15.1",
+    "sharp": "^0.33.5",
+    "vue": "latest",
+    "vue-router": "latest",
+    "vue3-dropzone": "^2.2.1"
   },
   "devDependencies": {
-    "@types/file-saver": "^2.0.5",
-    "@types/node": "^20.2.5",
-    "@types/react": "^18.2.9",
-    "@types/react-dom": "^18.2.4",
-    "@typescript-eslint/eslint-plugin": "^5.59.9",
-    "@typescript-eslint/parser": "^5.59.9",
-    "autoprefixer": "^10.4.14",
-    "eslint": "^8.42.0",
-    "eslint-config-next": "13.4.4",
-    "eslint-config-prettier": "^8.8.0",
-    "eslint-plugin-prettier": "^4.2.1",
-    "eslint-plugin-tailwindcss": "^3.12.1",
-    "postcss": "^8.4.24",
-    "prettier": "^2.8.8",
-    "prettier-plugin-tailwindcss": "^0.3.0",
-    "standard": "^17.1.0",
-    "tailwindcss": "^3.3.2",
-    "typescript": "^5.1.3"
+    "@nuxtjs/eslint-module": "^4.1.0",
+    "@typescript-eslint/eslint-plugin": "^8.19.1",
+    "@typescript-eslint/parser": "^8.19.1",
+    "eslint": "^9.17.0",
+    "eslint-plugin-nuxt": "^4.0.0",
+    "typescript": "^5.7.2"
   }
 }
